pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "femsolver-tools"
version = "0.1.0"
description = "Repository tooling for the femsolver project."
requires-python = ">=3.11"
dependencies = ["PyYAML>=6.0"]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.4", "gcovr>=8.0"]

[tool.pytest.ini_options]
testpaths = ["tests/python"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
